% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodExpMagpieMarkup.R
\name{FoodExpMagpieMarkup}
\alias{FoodExpMagpieMarkup}
\title{FoodExpMagpieMarkup}
\usage{
FoodExpMagpieMarkup(
  gdx,
  level = "reg",
  type = "consumer",
  prodAggr = FALSE,
  afterShock = FALSE
)
}
\arguments{
\item{gdx}{gdx file as input for mag base prices}

\item{level}{reg or regglo}

\item{type}{"consumer" (marked up) or "producer" price output}

\item{prodAggr}{TRUE for aggregating across alll products}

\item{afterShock}{TRUE or FALSE for kcal consumed}
}
\value{
dataframe or magclass object of consumer food expenditures
}
\description{
calculate markups
}
\author{
David M Chen
}
