% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readICP2017raw.R
\name{readICP2017raw}
\alias{readICP2017raw}
\title{readICP2017raw}
\usage{
readICP2017raw()
}
\value{
dataframe of ICP expenditures
}
\description{
read ICP data, not share-able data, contained in a local readSource
}
\author{
David M Chen
}
